remove tokencontext
